Move TestCLI to separate file
